Rename bindings/ to codegen/
