+++ codegen/generate_class.py
