+++ core/scripting/ScriptEnvironment.cc
@@ -5,7 +5,7 @@
 
 #include <vector>
 
-#include "bindings/script_linker.h"
+#include "codegen/script_linker.h"
 #include "core/assets/ScriptAsset.h"
 #include "core/components/ScriptComponent.h"
 #include "core/components/TransformComponent.h"
